functions/genimage/requirements.txt
functions-framework==3.*
google-cloud-datastore
google-cloud-storage
openai